% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_add_diagonals_to_grid.R
\name{add_diagonals_to_grid}
\alias{add_diagonals_to_grid}
\title{Create a \code{multilinestring} sf object for the diagonal and off-diagonal lines
for each grid cell}
\usage{
add_diagonals_to_grid(sf_object = NULL)
}
\arguments{
\item{sf_object}{An sf object (tibble) representing a spatial grid. The
function expects this object to have a geometry column with polygons
representing grid cells. If \code{NULL}, the function will stop with an error
message.}
}
\value{
An \code{sf} object where each row corresponds to a grid cell from the
input, represented by a \code{multilinestring} geometry of its diagonal and
off-diagonal lines. The returned object retains the coordinate reference
system (CRS) of the input.
}
\description{
This function takes an \code{sf} object representing a grid and creates a new sf
object where each grid cell is represented by a \code{multilinestring} geometry
consisting of its diagonal and off-diagonal lines.
}
\note{
The function requires the \code{sf}, \code{dplyr}, \code{purrr}, \code{tibble}, and \code{tidyr}
packages to be installed and loaded.
}
\examples{
ecokit::load_packages(package_list = c("dplyr", "sf", "raster", "ggplot2"))

Grid <- raster::raster(nrows = 10, ncols = 10, xmn = 0, xmx = 10,
                       ymn = 0, ymx = 10, crs = 4326) \%>\%
  setNames("Grid") \%>\%
  raster::setValues(1) \%>\%
  raster::rasterToPolygons() \%>\%
  sf::st_as_sf()

ggplot2::ggplot() +
  ggplot2::geom_sf(Grid, mapping = ggplot2::aes(), color = "black",
                   linewidth = 0.5, fill = "transparent") +
  ggplot2::scale_x_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::scale_y_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::theme_minimal()

Grid_X <- add_diagonals_to_grid(sf_object = Grid)

ggplot2::ggplot() +
  ggplot2::geom_sf(Grid, mapping = ggplot2::aes(), color = "black",
                   linewidth = 0.5, fill = "transparent") +
  ggplot2::geom_sf(Grid_X, mapping = ggplot2::aes(), color = "red",
                   linewidth = 0.5, inherit.aes = TRUE) +
  ggplot2::scale_x_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::scale_y_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::theme_minimal()
}
\seealso{
add_cross_to_grid
}
\author{
Ahmed El-Gabbas
}
