% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_all_objects_sizes.R
\name{all_objects_sizes}
\alias{all_objects_sizes}
\title{Size of objects in memory}
\usage{
all_objects_sizes(
  greater_than = 0L,
  in_function = FALSE,
  n_decimals = 2L,
  n_objects = Inf
)
}
\arguments{
\item{greater_than}{Numeric. Size threshold in MB. Only objects larger than
this value will be shown. Default is 0, which means all objects will be
shown. \code{greater_than} must be a non-negative number.}

\item{in_function}{Logical. This controls the scope of the function. It
indicates whether the execution is done inside or outside of a function.
Defaults to \code{FALSE} to show sizes of objects in the global environment. If
set to \code{TRUE}, sizes of objects in the function are returned.}

\item{n_decimals}{Integer; representing the number of decimal places to
show in the \code{size_mb} column. Defaults to 2.}

\item{n_objects}{Number of objects to show. Defaults to \code{Inf} meaning show
all available objects.}
}
\value{
The function prints a tibble containing the variables' names, their
sizes in MB, and their percentage of the total size of all variables. If no
objects meet the criteria, a message is printed instead. Output is sorted
in descending order of the size of the objects. The function also prints
the total size of all variables and the number of objects that were
examined.
}
\description{
This function calculates the size of objects in the global environment of R
using \link[lobstr:obj_size]{lobstr::obj_size} and prints a summary of objects that are greater
than a specified size threshold. It is useful for memory management and
identifying large objects in the workspace.
}
\examples{
AA1 <<- rep(seq_len(1000), 10000)
AA2 <<- rep(seq_len(1000), 100)

# ----------------------------------------------------
# All objects in memory
# ----------------------------------------------------

all_objects_sizes()


# ----------------------------------------------------
# Objects larger than 1 MB
# ----------------------------------------------------

all_objects_sizes(greater_than = 1)


# ----------------------------------------------------
# Objects larger than 50 MB
# ----------------------------------------------------

all_objects_sizes(greater_than = 50)


# ----------------------------------------------------
# When called with another function, it shows the objects only available
# within the function
# ----------------------------------------------------

TestFun <- function(XX = 10) {
  Y <- 20
  C <- matrix(data = seq_len(10000), nrow = 100, ncol = 100)
  all_objects_sizes(in_function = TRUE)
}

TestFun()

TestFun(XX = "TEST")
}
\author{
Ahmed El-Gabbas
}
