% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_boundary_to_wkt.R
\name{boundary_to_wkt}
\alias{boundary_to_wkt}
\title{Determine the boundaries of the requested GBIF data}
\usage{
boundary_to_wkt(left = NULL, right = NULL, bottom = NULL, top = NULL)
}
\arguments{
\item{left, right, bottom, top}{Numeric, the left, right, bottom, and top
boundary of the area.}
}
\value{
A character string representing the WKT of the polygon that outlines
the specified boundaries.
}
\description{
This function constructs a Well-Known Text (WKT) string representing a
polygon that outlines the specified boundaries. It is used to define the area
of interest for downloading GBIF data through the \code{\link[rgbif:download_predicate_dsl]{rgbif::pred_within()}}
function.
}
\examples{
boundary_to_wkt(left = 20, right = 30, bottom = 40, top = 50)
}
\author{
Ahmed El-Gabbas
}
