% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cat_diff.R
\name{cat_diff}
\alias{cat_diff}
\title{Print time difference}
\usage{
cat_diff(
  init_time,
  chunk_text = "Session summary",
  prefix = "Completed in ",
  cat_info = FALSE,
  level = 0L,
  cat_timestamp = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{init_time}{\code{POSIXct}. The initial time from which the difference is
calculated.}

\item{chunk_text}{Character. The message printed as chunk info. Default
value: \verb{Session summary}. See: \link{info_chunk} for more information.}

\item{prefix}{Character. prefix to prepend to the printed time difference.
Defaults to "Completed in ".}

\item{cat_info}{Logical. If \code{TRUE}, prints a session summary using
\link{info_chunk} ("Session summary"). Defaults to \code{FALSE}.}

\item{level}{integer; the level at which the message will be printed. If e.g.
\code{level = 1L}, the following string will be printed at the beginning of the
message: "  >>  ". Default is \code{0}.}

\item{cat_timestamp}{logical; whether to include the time in the timestamp.
Default is \code{TRUE}. If \code{FALSE}, only the text is printed.}

\item{verbose}{logical; whether to print output to console. Default is
\code{TRUE}. If \code{FALSE}, the function does nothing. This is useful to suppress
the function output in certain contexts.}

\item{...}{Additional arguments for \link{cat_time}.}
}
\value{
The function is used for its side effect of printing to the console
and does not return any value.
}
\description{
This function calculates the time difference from a given initial time to the
current time and prints it with a specified prefix. Optionally, it can also
print a session summary.
}
\examples{
# basic usage
reference_time <- (lubridate::now() - lubridate::seconds(45))

cat_diff(reference_time)

# custom prefix text
cat_diff(reference_time, prefix = "Finished in ")

# level = 1
cat_diff(reference_time, prefix = "Finished in ", level = 1L)

# print date
cat_diff(reference_time, prefix = "Finished in ", cat_timestamp = TRUE)

# print date and time
cat_diff(reference_time, prefix = "Finished in ", cat_date = TRUE)

# show chunk info
cat_diff(reference_time, cat_info = TRUE, prefix = "Finished in ")

# custom chunk info text
cat_diff(
  reference_time, cat_info = TRUE, chunk_text = "Summary of task",
  prefix = "Finished in ")

# |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||

reference_time <- (lubridate::now() -
    (lubridate::minutes(50) + lubridate::seconds(45)))
cat_diff(reference_time)

reference_time <- (lubridate::now() - lubridate::minutes(50))
cat_diff(reference_time)

reference_time <- (lubridate::now() - lubridate::minutes(70))
cat_diff(reference_time)

reference_time <- (lubridate::now() - lubridate::hours(4))
cat_diff(reference_time)

reference_time <- lubridate::now() -
  (lubridate::hours(4) + lubridate::minutes(50) + lubridate::seconds(45))
cat_diff(reference_time)
}
\author{
Ahmed El-Gabbas
}
