% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_check_rstudio.R
\name{check_rstudio}
\alias{check_rstudio}
\title{Check if the installed RStudio version is up to date}
\usage{
check_rstudio()
}
\value{
Side effects include printing messages to the console regarding the
status of RStudio version.
}
\description{
This function checks the current installed version of RStudio against the
latest version available online. If the versions do not match, it suggests
updating RStudio.
}
\note{
This function requires internet access to check the latest version of
RStudio online. If called outside of RStudio, it will only fetch and
display the latest version without comparing. If posit has changed the
structure of their webpage and the online version can not be parsed, the
function reports this clearly and, when possible, still shows the installed
version.
}
\examples{
check_rstudio()
}
\author{
Ahmed El-Gabbas
}
