% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_check_tiff.R
\name{check_tiff}
\alias{check_tiff}
\title{Check if a tiff file corrupted}
\usage{
check_tiff(x = NULL, warning = TRUE)
}
\arguments{
\item{x}{Character; the file path of the tiff file to be checked. The
function will stop with an error if \code{x} is \code{NULL} or if the file does not
exist.}

\item{warning}{Logical. If \code{TRUE}, the function will issue a warning if the
file does not exist.}
}
\value{
Logical; returns \code{TRUE} if the TIFF file is not corrupted (i.e., the
file exists, can be described with a "Driver" in its metadata, has values,
and its data can be read without errors or warnings), and \code{FALSE}
otherwise.
}
\description{
This function checks if the provided tiff file is corrupted by attempting to
describe it using the \code{terra} package and searching for the presence of a
"Driver" string in the description, which indicates a valid tiff file. If the
string is found, the function returns \code{TRUE} and \code{FALSE} otherwise. The
function works also for reading netCDF files with the \code{terra} package.
}
\examples{
(f <- system.file("ex/elev.tif", package="terra"))

check_tiff(x = f)

# a temp file ends with .tif (not a valid tiff file)
(temp_file <- tempfile(fileext = ".tif"))
fs::file_create(temp_file)
check_tiff(x = temp_file)

# clean up
fs::file_delete(temp_file)
}
\author{
Ahmed El-Gabbas
}
