% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_clear_console.R
\name{clear_console}
\alias{clear_console}
\title{Clear the console}
\usage{
clear_console()
}
\value{
An invisible \code{NULL} to indicate the function has completed without
returning any meaningful value.
}
\description{
function clears the console in RStudio by sending a form feed character. If
not run in RStudio, it prints a message indicating the function is not
supported.
}
\note{
This function checks if it is being run in RStudio by examining the
\code{RSTUDIO} environment variable. If the function is not run in RStudio, it
will not clear the console and instead print a message.
}
\examples{
clear_console()
}
