% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_file_extension.R
\name{file_extension}
\alias{file_extension}
\title{Get the file extension from a file path}
\usage{
file_extension(path)
}
\arguments{
\item{path}{A character string representing the file path from which the file
extension is to be extracted. It must not be \code{NULL} and has to be a
character string.}
}
\value{
A character string representing the file extension of the given file
path. If the path does not have an extension, an empty string is returned.
}
\description{
This function extracts the file extension from a given file path. It first
checks if the input is not NULL and is a character string. If these
conditions are met, it then uses the \link[tools:fileutils]{tools::file_ext} function to extract
and return the file extension. The function does not check the existence of
the file or explicitly get the file type from its content. It merely guess
file extension from the file name.
}
\examples{
file_extension(path = "File.doc")

file_extension(path = "D:/File.doc")

file_extension(path = "File.1.doc")

file_extension(path = "D:/Files.All")

file_extension(path = "D:/Files.All/")

file_extension("example.txt") # returns "txt"

file_extension("archive.tar.gz") # returns "gz"
}
\seealso{
\code{\link[=file_type]{file_type()}}
}
\author{
Ahmed El-Gabbas
}
