% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_file_type.R
\name{file_type}
\alias{file_type}
\title{Determine the file type of a given file path}
\usage{
file_type(path)
}
\arguments{
\item{path}{Character. The path to the file whose type is to be determined.
The path must not be \code{NULL}, and the file must exist.}
}
\value{
A character string describing the file type.
}
\description{
This function uses the system's \code{file} command to determine the type of the
file specified by the \code{path} parameter. It returns a character string
describing the file type.
}
\note{
This function relies on the system's \code{file} command and therefore
might produce different outputs on different platforms.
}
\examples{
(f <- system.file("ex/elev.tif", package="terra"))

file_type(path = f)
}
\author{
Ahmed El-Gabbas
}
