% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Calculate the mode of a numeric vector}
\source{
The source of this function was taken from this
\href{https://www.tutorialspoint.com/r/r_mean_median_mode.htm}{link}.
}
\usage{
get_mode(x)
}
\arguments{
\item{x}{Numeric vector. It must not be \code{NULL} or empty.}
}
\value{
The mode of the vector as a single value. If the vector has a uniform
distribution (all values appear with the same frequency), the function
returns the first value encountered.
}
\description{
This function calculates the mode of a given numeric vector.
}
\examples{
get_mode(c(seq_len(10), 1, 1, 3, 3, 3, 3))

get_mode(c(1, 2, 2, 3, 4))

get_mode(c(1, 1, 2, 3, 3))
}
