% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_options.R
\name{get_option_with_default}
\alias{get_option_with_default}
\title{Retrieve Option Value with Function Argument Default Fallback}
\usage{
get_option_with_default(option_name, fun_name, arg_name)
}
\arguments{
\item{option_name}{Character. The name of the R option to retrieve (e.g.,
\code{"my_pkg_option"}).}

\item{fun_name}{Character. The name of the function, optionally qualified
with a package (e.g., \code{"my_fun"} or \code{"mypkg::my_fun"}).}

\item{arg_name}{Character. The name of the argument whose default value
should be used as a fallback.}
}
\value{
The value of the option if set, otherwise the default value of the
specified function argument. Returns \code{NULL} if the default is not
accessible (e.g., for primitive or non-standard functions).
}
\description{
This function returns the value of a specified R option if it is set;
otherwise, it falls back to the default value of a specified argument of a
given function. The function can be identified by name only or with package
qualification (e.g., \code{"pkg::fun"}). It supports arguments with default values
that are constants or quoted expressions.
}
\details{
\strong{Important:} This function only works for standard R functions whose
default argument values are accessible via \code{\link[=formals]{formals()}}. It does \strong{not} work
for primitive functions (such as \code{max}, \code{mean}), functions implemented in
C/C++, S4 methods, or functions whose defaults are not accessible
programmatically.
}
\examples{
# No option is called `.add_changed`
getOption(".add_changed")

# return the default value of the `.add` argument of [dplyr::group_by()]
get_option_with_default(".add_changed", "dplyr::group_by", ".add")

# Setting and retrieving the option
options(.add_changed = TRUE)
get_option_with_default(".add_changed", "dplyr::group_by", ".add")

# Removing the option, should fall back to default again
ecokit::remove_options(".add_changed")
get_option_with_default(".add_changed", "dplyr::group_by", ".add")

# Will return NULL for primitives:
get_option_with_default("base_max_na.rm", "base::max", "na.rm")

}
\author{
Ahmed El-Gabbas
}
