% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load or install multiple R packages}
\usage{
load_packages(
  ...,
  package_list = NULL,
  verbose = FALSE,
  install_missing = FALSE,
  n_cpus = getOption("Ncpus", 1L)
)
}
\arguments{
\item{...}{Character. Names of the packages to be loaded or installed.}

\item{package_list}{Character vector. An alternative or additional way to
specify package names as a vector.}

\item{verbose}{Logical. If \code{TRUE}, prints the names and versions of the
loaded packages. Defaults to \code{FALSE}.}

\item{install_missing}{Logical. If \code{TRUE}, missing packages are automatically
installed and then loaded. Defaults to \code{FALSE}.}

\item{n_cpus}{Integer. Number of CPUs to use for parallel installation of
packages. Defaults to the value of the \code{Ncpus} option. This is only valid
if \code{install_missing} is \code{TRUE}.}
}
\value{
This function is used for its side effects (loading/installing
packages) and does not return any value.
}
\description{
This function attempts to load multiple R packages specified by the user. If
a package is not installed, the function can optionally install it before
loading. It also provides an option to print the names and versions of the
loaded packages.
}
\details{
On HPC systems using network filesystems (e.g. Lustre) with \code{renv},
\code{\link[utils:installed.packages]{utils::installed.packages()}} can be slow due to high metadata I/O latency
across large library trees. This function mitigates that by calling
\code{\link[=installed.packages]{installed.packages()}} once with \code{fields = "Version"} only, caching
the result, and reading package versions from the cached matrix rather than
re-reading individual \code{DESCRIPTION} files via \code{packageDescription()}.
}
\examples{
# Currently loaded packages
(P1 <- ecokit::loaded_packages())

# Load tidyr
load_packages(tidyr, raster, ggplot2, nnet, verbose = TRUE)

# Loaded packages after implementing the function
(P2 <- ecokit::loaded_packages())

# Which packages were loaded?
setdiff(P2, P1)

# verbose = FALSE (default)
load_packages(tidyterra, verbose = FALSE)

# load already loaded packages
load_packages(tidyr, tidyterra, verbose = TRUE)

# non-existent package
load_packages("non_existent")
}
\author{
Ahmed El-Gabbas
}
