% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_get_sampling_effort.R
\name{mask_cumulative_pct}
\alias{mask_cumulative_pct}
\title{Mask raster to show top \% and bottom \% of cumulative sum}
\usage{
mask_cumulative_pct(rast, top_pct = 90L)
}
\arguments{
\item{rast}{A SpatRaster object with numeric values (e.g., counts per cell)}

\item{top_pct}{Numeric, percentage (0–100) for the top cumulative sum
(default: 90)}
}
\value{
A SpatRaster with three layers:
\itemize{
\item \code{top}: cells cumulatively accounting for the top \code{top_pct}\%
\item \code{bottom}: cells cumulatively accounting for the lowest \code{(100 - top_pct)}\%
\item \code{zero_observations}: cells with original value equal to zero
}
}
\description{
Creates two masked rasters:
\itemize{
\item \code{top}: cells cumulatively accounting for the top \code{top_pct} \% of total sum
\item \code{bottom}: cells cumulatively accounting for the remaining bottom
\code{(100 - top_pct)} \%
}
}
\examples{
require(terra)

# Sampling effort raster for birds (number of observations at 20 km
# resolution)
efforts_birds_all <- get_sampling_effort(
  group = "aves", metric = "n_obs", resolution = 20)
efforts_birds_all_r <- terra::rast(efforts_birds_all$local_path)

result <- mask_cumulative_pct(rast = efforts_birds_all_r, top_pct = 90)
result

# Areas contributing to the top 90\% of cumulative sum (log10 scale; computed
# on the global scale and cropped to USA)
result$top_90_percent_cumulative \%>\%
  terra::crop(terra::ext(-125, -66.5, 24.5, 49.5)) \%>\%
  log10() \%>\%
  plot()

# Areas contributing to the lowest 10\% of cumulative sum (log10 scale;
# computed on the global scale and cropped to USA)
result$lowest_10_percent_cumulative \%>\%
  terra::crop(terra::ext(-125, -66.5, 24.5, 49.5)) \%>\%
  log10() \%>\%
  plot()

# Areas with zero observations (computed on the global scale and cropped to
# USA)
result$zero_observations \%>\%
  terra::crop(terra::ext(-125, -66.5, 24.5, 49.5)) \%>\%
  plot()

}
