% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_maxent.R
\name{maxent_variable_importance}
\alias{maxent_variable_importance}
\title{Extract Variable Importance from a Maxent Model Object}
\usage{
maxent_variable_importance(model = NULL)
}
\arguments{
\item{model}{A fitted Maxent model object of class "MaxEnt" (from
\code{\link[dismo:maxent]{dismo::maxent()}}.}
}
\value{
A tibble with columns:
\describe{
\item{variable}{Variable name (character).}
\item{percent_contribution}{Percent contribution of the variable
(numeric).}
\item{permutation_importance}{Permutation importance of the variable
(numeric).}
}
}
\description{
This function extracts the percent contribution and permutation importance
for each variable from a fitted Maxent model (class "MaxEnt") as returned by
\code{\link[dismo:maxent]{dismo::maxent()}}. It returns a tibble with variables and their respective
importance metrics.
}
\examples{
require(ecokit)
ecokit::load_packages(fs, dismo, rJava, raster)

# fit a Maxent model
if (dismo::maxent(silent = TRUE)) {
  predictors <- list.files(
    path = fs::path(
      system.file(package = "dismo"), "ex"),
    pattern = "grd", full.names = TRUE) \%>\%
    raster::stack()

  occurence <- fs::path(
    system.file(package = "dismo"), "ex", "bradypus.csv") \%>\%
    read.table(header = TRUE, sep = ",") \%>\%
    dplyr::select(-1)
  # fit model, biome is a categorical variable
  me <- maxent(predictors, occurence, factors='biome')

  maxent_variable_importance(me)
}

}
\author{
Ahmed El-Gabbas
}
