% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_n_decimals.R
\name{n_decimals}
\alias{n_decimals}
\title{Number of decimal places in a numeric or character value}
\usage{
n_decimals(x = NULL)
}
\arguments{
\item{x}{Numeric or character vector representing numeric values.}
}
\value{
An integer vector of the same length as \code{x}, where each element
represents the number of decimal places in the corresponding input value.
}
\description{
This function calculates the number of decimal places in a numeric or
character value by counting all digits after the decimal point in the string
representation, including trailing zeros for characters. It is vectorized and
designed to work with numeric inputs or character strings representing
numbers, making it suitable for use with \code{dplyr::mutate}.
}
\examples{
# Character input with trailing zeros
n_decimals(c("1.35965", "65.5484900000", "0.11840000"))

# Numeric input with trailing zeros
n_decimals(c(1.35965, 65.5484900000, 0.11840000))

# Use with dplyr
library(dplyr)
mtcars \%>\%
  dplyr::select(wt) \%>\%
  dplyr::mutate(n_decimals = n_decimals(wt))
}
\author{
Ahmed El-Gabbas
}
