% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_os.R
\name{os}
\alias{os}
\title{Current operating system}
\usage{
os()
}
\value{
A character string representing the name of the operating system.
}
\description{
This function returns the name of the current operating system the R session
is running on.
}
\examples{
os()

}
\author{
Ahmed El-Gabbas
}
