% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_range_to_new_value.R
\name{range_to_new_value}
\alias{range_to_new_value}
\title{Changes values within a specified range, or greater than or less than a
specific value to a new value in a vector, data.frame, or raster}
\usage{
range_to_new_value(
  x = NULL,
  between = NULL,
  greater_than = NULL,
  less_than = NULL,
  new_value = NULL,
  invert = FALSE
)
}
\arguments{
\item{x}{A numeric \code{vector}, \code{data.frame}, \code{RasterLayer}, \code{SpatRaster}, or
\code{PackedSpatRaster} object whose values are to be modified.}

\item{between}{Numeric. A numeric vector of length 2 specifying the range of
values to be changed or kept. If specified, \code{greater_than} and \code{less_than}
are ignored.}

\item{greater_than, less_than}{Numeric. Threshold larger than or equal to/less
than or equal to which values in \code{x} will be changed to \code{new_value}. Only
applied if \code{between} is not specified.}

\item{new_value}{The new value to assign to the selected elements in \code{x}.}

\item{invert}{Logical. Whether to invert the selection specified by
\code{between}. If \code{TRUE}, values outside the specified range are changed to
\code{new_value}. Default is \code{FALSE}.}
}
\value{
The modified object \code{x} with values changed according to the
specified conditions.
}
\description{
This function modifies values in the input object \code{x} based on the specified
conditions. It can operate on vectors, data.frames, or RasterLayer objects.
The function allows for changing values within a specified range (\code{between}),
greater than or equals to (\code{greater_than}) or  less than or equals to
(\code{less_than}) a specified value to a new value (\code{new_value}). An option to
invert the selection is also available for ranges.
}
\examples{
ecokit::load_packages(dplyr, raster, terra, tibble, ggplot2, tidyr)

# ---------------------------------------------

# Vector

(VV <- seq_len(10))

range_to_new_value(x = VV, between = c(5, 8), new_value = NA)

range_to_new_value(x = VV, between = c(5, 8), new_value = NA, invert = TRUE)

# greater_than is ignored as `between` is specified
range_to_new_value(
   x = VV, between = c(5, 8), new_value = NA, greater_than = 4)
range_to_new_value(x = VV, new_value = NA, greater_than = 4)

range_to_new_value(x = VV, new_value = NA, less_than = 4)

# `invert` argument works only when `between` is specified
range_to_new_value(x = VV, new_value = NA, greater_than = 4, invert = TRUE)

# ---------------------------------------------

# tibble

iris2 <- iris \%>\%
  tibble::as_tibble() \%>\%
  dplyr::slice_head(n = 50) \%>\%
  dplyr::select(-Sepal.Length, -Petal.Length, -Petal.Width) \%>\%
  dplyr::arrange(Sepal.Width)

iris2 \%>\%
 dplyr::mutate(
   Sepal.Width.New = range_to_new_value(
      x = Sepal.Width, between = c(3, 3.5),
      new_value = NA, invert = FALSE),
   Sepal.Width.Rev = range_to_new_value(
      x = Sepal.Width, between = c(3, 3.5),
      new_value = NA, invert = TRUE)) \%>\%
 print(n = 50)

# ---------------------------------------------

# RasterLayer

grd_file <- system.file("external/test.grd", package = "raster")
R_raster <- raster::raster(grd_file)

# set the theme for ggplot2
ggplot2::theme_set(
  ggplot2::theme_minimal() +
  ggplot2::theme(
    legend.position = "right",
    strip.text = ggplot2::element_text(size = 16),
    legend.title = ggplot2::element_blank(),
    axis.title = ggplot2::element_blank(),
    axis.text = ggplot2::element_blank()))

# Convert values less than 500 to NA
R_raster2 <- range_to_new_value(
  x = R_raster, less_than = 500, new_value = NA)
# Convert values greater than 600 to NA
R_raster3 <- range_to_new_value(
   x = R_raster, greater_than = 600, new_value = NA)
(R_rasters <- raster::stack(R_raster, R_raster2, R_raster3))


as.data.frame(R_rasters, xy = TRUE, na.rm = FALSE) \%>\%
  stats::setNames(c("x", "y", "R_raster", "R_raster2", "R_raster3")) \%>\%
  tidyr::pivot_longer(
    cols = -c("x", "y"), names_to = "layer", values_to = "value") \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_tile(mapping = ggplot2::aes(x = x, y = y, fill = value)) +
  ggplot2::facet_grid(~layer) +
  ggplot2::scale_fill_gradientn(
    colours = c("blue", "green", "yellow", "red"),
    na.value = "transparent") +
  ggplot2::labs(title = NULL, x = NULL, y = NULL) +
  ggplot2::coord_cartesian(expand = FALSE, clip = "off")

# ---------------------------------------------

# SpatRaster

R_terra <- terra::rast(grd_file)
R_terra2 <- range_to_new_value(x = R_terra, less_than = 500, new_value = NA)
R_terra3 <- range_to_new_value(
    x = R_terra, greater_than = 600, new_value = NA)
(R_terras <- c(R_terra, R_terra2, R_terra3))

as.data.frame(R_terras, xy = TRUE, na.rm = FALSE) \%>\%
  stats::setNames(c("x", "y", "R_terra", "R_terra2", "R_terra3")) \%>\%
  tidyr::pivot_longer(
    cols = -c("x", "y"), names_to = "layer", values_to = "value") \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_tile(mapping = ggplot2::aes(x = x, y = y, fill = value)) +
  ggplot2::facet_grid(~layer) +
  ggplot2::scale_fill_gradientn(
    colours = c("blue", "green", "yellow", "red"),
    na.value = "transparent") +
  ggplot2::labs(title = NULL, x = NULL, y = NULL) +
  ggplot2::coord_cartesian(expand = FALSE, clip = "off")
}
\author{
Ahmed El-Gabbas
}
