% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_raster_to_PA.R
\name{raster_to_PA}
\alias{raster_to_PA}
\title{Convert raster map into binary (1/0)}
\usage{
raster_to_PA(raster = NULL, NA_to_0 = TRUE, zero_to_NA = FALSE)
}
\arguments{
\item{raster}{The input raster map. It must be of class \code{PackedSpatRaster},
\code{RasterLayer}, or \code{SpatRaster}. This parameter cannot be NULL.}

\item{NA_to_0}{A logical value indicating whether NA values should be
converted to 0. Defaults to \code{TRUE}.}

\item{zero_to_NA}{A logical value indicating whether 0 values should be
converted to NA. Defaults to \code{FALSE}.}
}
\value{
A raster map where values have been converted according to the
specified parameters. This object is of the same class as the input object.
}
\description{
This function converts raster values into a binary format where positive
values are set to 1 (presence) and zeros remain 0 (absence). Additionally, it
allows for the conversion of NA values to 0, and/or 0 values to NA, based on
the user's choice.
}
\examples{
ecokit::load_packages(
   package_list = c("dplyr", "raster", "ggplot2", "tidyterra"))

r <- r2 <- raster::raster(
  system.file("external/test.grd", package = "raster"))

# change some values to 0
r[5000:6000] <- 0
r <- raster::mask(r, r2)

ggplot2::ggplot() +
  tidyterra::geom_spatraster(data = terra::rast(r), maxcell = Inf) +
  ggplot2::theme_minimal()

r_2 <- raster::stack(
  # 1) NA replaced with 0
  raster_to_PA(raster = r),
  # 2) NA is kept as NA
  raster_to_PA(raster = r, NA_to_0 = FALSE),
  # 3) 0 replaced with NA
  raster_to_PA(raster = raster_to_PA(r), zero_to_NA = TRUE))

ggplot2::ggplot() +
  tidyterra::geom_spatraster(
    data = terra::as.factor(terra::rast(r_2)), maxcell = Inf) +
  ggplot2::facet_wrap(~lyr) +
  ggplot2::scale_fill_manual(values = c("grey30", "red", "blue"),
    na.value = "transparent") +
  ggplot2::theme_minimal()
}
\author{
Ahmed El-Gabbas
}
