% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_save_as.R
\name{save_as}
\alias{save_as}
\title{Save an object to a file with a new name}
\usage{
save_as(
  object = NULL,
  object_name = NULL,
  out_path = NULL,
  n_threads = 1L,
  feather_compression = "zstd",
  ...
)
}
\arguments{
\item{object}{The input object to be saved. This can be an actual R object or
a character string representing the name of an object. If input object is a
\code{SpatRaster}, it will be converted to a \code{PackedSpatRaster} before saving.}

\item{object_name}{Character. The new name for the saved \code{RData} object. This
name is used when the object is loaded back into R. Default is \code{NULL}. This
is required when saving \code{RData} files.}

\item{out_path}{Character. File path (ends with either \verb{*.RData}, \verb{*.qs2},
\code{feather}, and \code{rds}) where the object be saved. This includes the
directory and the file name.}

\item{n_threads}{Numeric. Number of threads to use when compressing data. See
\link[qs2:qs_save]{qs2::qs_save}.}

\item{feather_compression}{Character. The compression algorithm to use when
saving the object in the \code{feather} format. The default is "zstd". See
\link[arrow:write_feather]{arrow::write_feather}.}

\item{...}{Additional arguments to be passed to the respective save
functions. \link[base:save]{base::save} for \code{RData} files; \link[qs2:qs_save]{qs2::qs_save} for \code{qs2} files;
\link[arrow:write_feather]{arrow::write_feather} for \code{feather} files; and \link[base:saveRDS]{base::saveRDS} for \code{rds}
files.}
}
\value{
The function does not return a value but saves an object to the
specified file path.
}
\description{
This function saves an R object to a specified file path with a potentially
new object name. It is useful for renaming objects during the save process.
The function supports saving objects in \code{RData}, \code{qs2}, \code{feather}, and \code{rds}
formats. The format is determined by the extension of the file path
(case-insensitive).
}
\examples{
load_packages(fs, tibble)

temp_dir <- fs::path_temp("save_as")
fs::dir_create(temp_dir)
out_file <- fs::path(temp_dir, "iris2.RData")
list.files(temp_dir)

# save iris data as `iris2.RData` with `iris2` object name
save_as(
  object = tibble::tibble(iris), object_name = "iris2", out_path = out_file)

list.files(temp_dir, pattern = "^.+.RData")

# load the object to global environment. The data is loaded as `iris2`
(loaded_name <- load(out_file))

ecokit::load_as(out_file)

# clean up
fs::file_delete(out_file)
}
\author{
Ahmed El-Gabbas
}
