% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_set_raster_CRS.R
\name{set_raster_CRS}
\alias{set_raster_CRS}
\title{sets CRS for a SpatRaster}
\usage{
set_raster_CRS(raster, CRS = "epsg:3035")
}
\arguments{
\item{raster}{A SpatRaster object whose CRS needs to be set.}

\item{CRS}{Character. CRS value to be set, default is "epsg:3035".}
}
\value{
The SpatRaster object with the updated CRS.
}
\description{
This function sets the coordinate reference system (CRS) for a SpatRaster
object using the specified EPSG code. This is a wrapper function for
\code{terra::crs(raster) <- CRS} but allowing to set the CRS in the pipe.
}
\author{
Ahmed El-Gabbas
}
