% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_split_vector.R
\name{split_vector}
\alias{split_vector}
\title{Split a vector into smaller chunks}
\usage{
split_vector(vector = NULL, n_splits = NULL, prefix = "Chunk")
}
\arguments{
\item{vector}{A numeric or character vector that you want to split.}

\item{n_splits}{Integer. Number of chunks to split the vector into. It must
not exceed the length of the vector.}

\item{prefix}{Character. prefix for the names of the chunks in the returned
list. Defaults to \code{"Chunk"}.}
}
\value{
A list of vectors, where each vector represents a chunk of the
original vector. The names of the list elements are generated using the
specified prefix followed by an underscore and the chunk number.
}
\description{
This function divides a given vector into a specified number of smaller
chunks. It is useful for partitioning data into more manageable pieces or for
parallel processing tasks.
}
\examples{
split_vector(vector = seq_len(100), n_splits = 3)

# -------------------------------------------

split_vector(vector = seq_len(100), n_splits = 2, prefix = "T")

}
\author{
Ahmed El-Gabbas
}
