% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_system_command.R
\name{system_command}
\alias{system_command}
\title{Run a system command in a cross-platform manner}
\usage{
system_command(command, r_object = TRUE, ...)
}
\arguments{
\item{command}{Character. The bash command to be executed.}

\item{r_object}{Logical. Whether to capture the output of the command as an R
object. If \code{TRUE} (Default), the output is captured; if \code{FALSE}, the output
is printed to the console.}

\item{...}{Additional arguments passed to either \code{shell} or \code{system}
function, depending on the operating system.}
}
\value{
Depending on the value of \code{r_object}, either the output of the
executed command as an R object or \code{NULL} if \code{r_object} is \code{FALSE} and the
output is printed to the console.
}
\description{
This function executes a system command, using either \code{shell} on Windows or
\code{system} on Linux. It allows the output of the command to be captured into an
R object.
}
\examples{
# print working directory
system_command("pwd")
system_command("pwd", r_object = FALSE)

# first 5 files on the working directory
(A <- system_command("ls | head -n 5"))

B <- system_command("ls | head -n 5", r_object = FALSE)
B
}
\author{
Ahmed El-Gabbas
}
