% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_used_packages.R
\name{used_packages}
\alias{used_packages}
\title{Extract package names used with :: in an R script}
\usage{
used_packages(file_path)
}
\arguments{
\item{file_path}{Character string specifying the path to the R script file.}
}
\value{
A character vector of unique package names used with \code{::}. Returns
\code{character(0)} if none are found.
}
\description{
Reads an R script file and extracts unique package names used with the \code{::}
operator (e.g., \code{dplyr} from \code{dplyr::arrange}). Ignores entire lines that are
comments (starting with \verb{#}, ignoring whitespace) and text after \verb{#} within
lines.
}
\examples{
# Example with a script from GitHub
dplyr_select_url <- paste0(
  "https://raw.githubusercontent.com/elgabbas/ecokit/",
  "refs/heads/main/R/spat_split_raster.R")
example_script <- fs::file_temp("Example_script_", ext = "R")
download.file(dplyr_select_url, destfile = example_script, quiet = TRUE)

used_packages(example_script)

# cleanup
fs::file_delete(example_script)
}
