% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_line.R
\name{add_line}
\alias{add_line}
\title{Add a horizontal or vertical line to the current plot}
\source{
The source code of this function was taken from this
\href{https://stackoverflow.com/questions/27800307/}{stackoverflow} question.
}
\usage{
add_line(at = NULL, outer = FALSE, horizontal = TRUE, ...)
}
\arguments{
\item{at}{Numeric; the relative location of where the line should be plotted.
Cannot be \code{NULL}.}

\item{outer}{Logical; if \code{TRUE}, the line is plotted outside of the plotting
area. Default is \code{FALSE}.}

\item{horizontal}{Logical; if \code{TRUE} (default), a horizontal line is added.
If \code{FALSE}, a vertical line is added.}

\item{...}{Additional graphical parameters passed to \link[graphics:abline]{graphics::abline}.}
}
\value{
Invisible; the function is called for its side effect of drawing on
the current plot.
}
\description{
Add a horizontal or vertical line to the current plot
}
\examples{
# Horizontal line
par(oma = c(1, 1, 1, 1), mar = c(3, 3, 1, 1))
plot(seq_len(100))

# add horizontal line
add_line(at = 0.25)
# add horizontal line that extends outside the plot area
add_line(at = 0.5, outer = TRUE)
# the same line but with a different line width and colour
add_line(at = 0.75, outer = TRUE, lwd = 3, col = "red")

# ---------------------------------------------

plot(seq_len(100))
# add vertical line
add_line(horizontal = FALSE, at = 0.25)
# add vertical line that extends outside the plot area
add_line(horizontal = FALSE, at = 0.5, outer = TRUE)
# the same line but with a different line width and colour
add_line(horizontal = FALSE, at = 0.75, outer = TRUE, lwd = 3, col = "red")
}
