% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_add_missing_columns.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns to a data frame with specified fill values}
\usage{
add_missing_columns(data, fill_value = NA_character_, ...)
}
\arguments{
\item{data}{A data frame to which missing columns will be added. This
parameter cannot be \code{NULL}.}

\item{fill_value}{The value to fill the missing columns with. This parameter
defaults to \code{NA_character_}, but can be changed to any scalar value as
required.}

\item{...}{Column names as character strings.}
}
\value{
a data frame with the missing columns added, if any were missing.
}
\description{
This function checks a data frame for missing columns specified by the user.
If any are missing, it adds these columns to the data frame, filling them
with a specified value.
}
\examples{
load_packages(dplyr, tibble)

# example data
(mtcars2 <- dplyr::select(mtcars, seq_len(3)) \%>\%
  head() \%>\%
  tibble::as_tibble())

mtcars2 \%>\%
 add_missing_columns(fill_value = NA_character_, A, B, C) \%>\%
 add_missing_columns(fill_value = as.integer(10), D)

AddCols <- c("Add1", "Add2")
mtcars2 \%>\%
 add_missing_columns(fill_value = NA_real_, AddCols)
}
\author{
Ahmed El-Gabbas
}
