% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_assign_env_vars.R
\name{assign_env_vars}
\alias{assign_env_vars}
\title{Assign environment variables from a .env file}
\usage{
assign_env_vars(env_file = ".env", env_variables_data = NULL)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{env_variables_data}{\code{data.frame}. A data frame or tibble containing the
columns \code{var_name}, \code{value}, \code{check_dir}, and \code{check_file}. Each row
specifies an environment variable, the name to assign it to, and whether to
check if it is a directory or file. This structure allows for additional
validation checks on the variables being imported.}
}
\value{
This function is used for its side effects of setting environment
variables in the R environment. It assigns each variable from the \code{.env}
file to the R environment with the name specified in the
\code{env_variables_data} data frame.
}
\description{
This function reads environment variables from a specified \code{.env} file and
assigns them to variables in the R environment based on a data frame
containing variable names, values, and checks for directories and files. It
is designed to facilitate the management of environment variables in a
structured and reproducible manner.
}
\author{
Ahmed El-Gabbas
}
