% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bash_variables.R
\name{bash_variables}
\alias{bash_variables}
\title{Read command line arguments passed to an R script}
\usage{
bash_variables()
}
\value{
This function does not return anything but has the side effect of
assigning variables into the global environment.
}
\description{
This function reads command line arguments passed to an R script executed via
the bash command line and assigns them into the global environment of the R
session. This allows for dynamic passing of variables from a bash script to
an R script.
}
\author{
Ahmed El-Gabbas
}
