% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cc.R
\name{cc}
\alias{cc}
\title{Concatenate without quotes}
\usage{
cc(...)
}
\arguments{
\item{...}{strings to be concatenated. Note that numeric values should be
converted to strings before being passed.}
}
\value{
A character string representing the concatenated values of the input
expressions.
}
\description{
This function takes one or more expressions and concatenates them into a
single string without quotes. It is particularly useful for creating strings
from variable names or expressions without including the usual quotes.
}
\examples{
cc(A, B, C)

# -------------------------------------------

cc(A, B, "A and B")

# -------------------------------------------

# this does not work
try(cc(A, B, "A and B", 10))

# this works
cc(A, B, "A and B", "10")
}
\author{
Ahmed El-Gabbas
}
