% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_check_data.R
\name{check_data}
\alias{check_data}
\alias{check_rdata}
\alias{check_qs}
\alias{check_rds}
\alias{check_feather}
\title{Check the integrity of \code{RData} / \code{qs2} / \code{rds} / \code{feather} files}
\usage{
check_data(file, warning = TRUE, n_threads = 5)

check_rdata(file, warning = TRUE)

check_qs(file, warning = TRUE, n_threads = 5)

check_rds(file, warning = TRUE)

check_feather(file, warning = TRUE)
}
\arguments{
\item{file}{Character. The path to the file to be checked. Cannot be empty.}

\item{warning}{Logical. If \code{TRUE} (default), warnings will be printed if the
file does not exist.}

\item{n_threads}{Integer. The number of threads to use when reading qs2
files. Default is 5.}
}
\value{
Logical. \code{TRUE} if the file is valid, otherwise \code{FALSE}.
}
\description{
These functions validate a given file by checking its extension and
attempting to load its contents. A file is considered valid if it loads
successfully and contains a non-null object, returning \code{TRUE}. Otherwise, it
returns \code{FALSE}.
}
\details{
The \code{check_data()} function determines the file type based on its
extension. If the extension is unrecognised, it returns \code{FALSE}. Supported
file types:
\itemize{
\item \strong{RData}: Checked with \code{check_rdata()}, read using \link{load_as}
\item \strong{qs2}: Checked with \code{check_qs()}, read using \link[qs2:qs_read]{qs2::qs_read}
\item \strong{rds}: Checked with \code{check_rds()}, read using \link{readRDS}
\item \strong{feather}: Checked with \code{check_feather()}, read using
\link[arrow:read_feather]{arrow::read_feather}
}
}
\author{
Ahmed El-Gabbas
}
