% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_check_quarto.R
\name{check_quarto}
\alias{check_quarto}
\title{Check if the installed Quarto version is up to date}
\usage{
check_quarto(pre_release = FALSE)
}
\arguments{
\item{pre_release}{Logical. Whether to check for pre-release versions.
Default is \code{FALSE}.}
}
\value{
A message indicating whether the installed Quarto version is up to
date or suggesting an update if it is not.
}
\description{
This function compares the installed Quarto version on the user's system with
the latest version available online. If the versions differ, it suggests the
user to update Quarto. It uses web scraping to find the latest version
available on the Quarto GitHub releases page and the system command to find
the installed version.
}
\examples{
check_quarto()

check_quarto(pre_release = TRUE)
}
\author{
Ahmed El-Gabbas
}
