% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_check_stack_in_memory.R
\name{check_stack_in_memory}
\alias{check_stack_in_memory}
\title{Check if a raster stack reads from disk or memory}
\usage{
check_stack_in_memory(stack = NULL)
}
\arguments{
\item{stack}{A RasterStack object. If \code{NULL} or not a RasterStack, the
function will stop with an error.}
}
\value{
No return value, but prints messages to the console.
}
\description{
This function checks whether the layers of a RasterStack object are stored in
memory or read from disk.  It prints messages indicating whether all layers
are in memory, all layers are on disk, or a mix of both. If there's a mix, it
specifies which layers are on disk.
}
\examples{
library(raster)
logo <- raster(system.file("external/rlogo.grd", package = "raster"))
logo@data@inmemory
logo@data@fromdisk
logo@file@name

# -------------------------------------------

# A raster stack reading from files
ST2 <- raster::stack(logo, logo)
check_stack_in_memory(ST2)
c(ST2[[1]]@data@inmemory, ST2[[2]]@data@inmemory)
c(ST2[[1]]@data@fromdisk, ST2[[2]]@data@fromdisk)
c(ST2[[1]]@file@name, ST2[[2]]@file@name)

# -------------------------------------------

logo2 <- raster::readAll(logo)
ST3 <- raster::stack(logo, logo2)
check_stack_in_memory(ST3)
c(ST3[[1]]@data@inmemory, ST3[[2]]@data@inmemory)
c(ST3[[1]]@data@fromdisk, ST3[[2]]@data@fromdisk)
c(ST3[[1]]@file@name, ST3[[2]]@file@name)
}
\author{
Ahmed El-Gabbas
}
