% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_check_url.R
\name{check_url}
\alias{check_url}
\title{Check the validity of URLs}
\source{
The source code of this function was taken from this
\href{https://stackoverflow.com/q/52911812}{stackoverflow} discussion.
}
\usage{
check_url(url = NULL, timeout = 2L, all_okay = TRUE)
}
\arguments{
\item{url}{Character. The URLs to be checked.}

\item{timeout}{Numeric. Timeout in seconds for the connection attempt.
Default is 2 seconds.}

\item{all_okay}{Logical. If \code{TRUE} (default), returns a single logical output
indicating the validity of all URLs; if \code{FALSE}, returns a logical vector
for each URL.}
}
\value{
Logical: \code{TRUE} if all checks pass; \code{FALSE} otherwise.
}
\description{
This function opens a connection to the specified URLs to check their
validity.
}
\examples{
urls <- c(
     "http://www.amazon.com", "http://this.isafakelink.biz",
     "https://stackoverflow.com", "https://stackoverflow505.com")

check_url(urls)

check_url(urls, all_okay = FALSE)
}
