% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecokit-package.R
\docType{package}
\name{ecokit-package}
\alias{ecokit}
\alias{ecokit-package}
\title{ecokit: Tools for Ecological and General Utilities}
\description{
A collection of utility functions designed for ecologists and general R users, providing tools for data manipulation, statistical analysis, and spatial data handling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elgabbas/ecokit}
  \item \url{https://elgabbas.github.io/ecokit/}
  \item Report bugs at \url{https://github.com/elgabbas/ecokit/issues}
}

}
\author{
\strong{Maintainer}: Ahmed El-Gabbas \email{elgabbas@outlook.com}

}
\keyword{internal}
