% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function_argumentss.R
\name{function_arguments}
\alias{function_arguments}
\title{Print function Arguments}
\usage{
function_arguments(function_name, assign = FALSE, package = NULL)
}
\arguments{
\item{function_name}{A function whose arguments you want to print. Must be a
valid R function.}

\item{assign}{Logical. Whether to assign the arguments as variables in the
global environment. Defaults to \code{FALSE}.}

\item{package}{Character. Name of the R package to be loaded before
processing the function. Default is \code{NULL}.}
}
\value{
The function prints the formatted arguments to the console. If
\code{assign} is TRUE, it also assigns arguments to the global environment.
}
\description{
This function takes another function as input and prints its arguments in the
format \code{ArgumentName = DefaultValue}. Each argument is printed on a new line.
The function can optionally assign the formatted arguments to the global
environment and can load a specified package before processing.
}
\examples{
formals(stats::setNames)

function_arguments(stats::setNames)
}
\author{
Ahmed El-Gabbas
}
