% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info_chunk.R
\name{info_chunk}
\alias{info_chunk}
\title{Print Information chunk with time stamp}
\usage{
info_chunk(
  message = "",
  cat_date = TRUE,
  sep_lines_before = 0L,
  sep_lines_after = 1L,
  cat_bold = FALSE,
  cat_red = FALSE,
  cat_timestamp = FALSE,
  level = 0L,
  msg_n_lines = 1L,
  info_lines_before = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{message}{Character. The main message to be timestamped. This parameter
is mandatory and cannot be \code{NULL} or empty.}

\item{cat_date}{Logical. Whether to include the date in the timestamp.
Default is \code{FALSE}, meaning only the time is printed. See
\link{cat_time}.}

\item{sep_lines_before, sep_lines_after}{Integer. Number of extra empty lines
to print before and after the separator lines. See \link{cat_sep} for
more details.}

\item{cat_bold}{logical; whether to print the text in bold. Default is
\code{FALSE}.}

\item{cat_red}{logical; whether to print the text in red. Default is \code{FALSE}.}

\item{cat_timestamp}{Logical. Whether to include the time in the timestamp.
Default is \code{FALSE}.}

\item{level}{integer; the level at which the message will be printed. If e.g.
\code{level = 1L}, the following string will be printed at the beginning of the
message: "   >>>   ". Default is \code{0}.}

\item{msg_n_lines}{integer; the number of newline characters to print after
the message. Default is 1.}

\item{info_lines_before}{Integer. Number of extra empty lines to print before
the message. Default is \code{0L}.}

\item{verbose}{logical; whether to print output to console. Default is
\code{TRUE}. If \code{FALSE}, the function does nothing. This is useful to suppress
the function output in certain contexts.}

\item{...}{Additional arguments passed to \link{cat_sep} for customizing
the separators.}
}
\value{
The function does not return any value but prints the message and
separators to the console.
}
\description{
This function prints a formatted message with a timestamp, surrounded by
separators for better readability in console outputs or logs.
}
\examples{
info_chunk(message = "Started")

info_chunk(message = "finished", line_char = "*", line_char_rep = 60)

info_chunk(message = "Started", cat_bold =  TRUE, cat_red = TRUE)

}
\author{
Ahmed El-Gabbas
}
