% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_list_to_rdata.R
\name{list_to_rdata}
\alias{list_to_rdata}
\title{Split list items into separate \code{.RData} files}
\usage{
list_to_rdata(list, prefix = "", directory = getwd(), overwrite = FALSE)
}
\arguments{
\item{list}{A named list object to be split into separate \code{.RData} files.}

\item{prefix}{Character. Prefix to each filename. If empty (default), no
prefix is added.}

\item{directory}{The directory where the \code{.RData} files will be saved.
Defaults to the current working directory.}

\item{overwrite}{A logical indicating whether to overwrite existing files.
Defaults to \code{FALSE}, in which case files that already exist will not be
overwritten, and a message will be printed for each such file.}
}
\value{
The function is called for its side effect of saving files and does
not return a value.
}
\description{
This function takes a named list and saves each element of the list as a
separate \code{.RData} file. The names of the list elements are used as the base
for the filenames, optionally prefixed. Files are saved in the specified
directory, with an option to overwrite existing files.
}
\examples{
load_packages(dplyr, fs)

# split iris data by species name
iris2 <- iris \%>\%
  tibble::tibble() \%>\%
  split(~Species)

str(iris2, 1)

# save each species as a separate RData file
temp_dir <- fs::path_temp("list_to_rdata")
fs::dir_create(temp_dir)
list.files(temp_dir)

list_to_rdata(list = iris2, directory = temp_dir)
list.files(temp_dir)

# loading data
setosa <- load_as(fs::path(temp_dir, "setosa.RData"))
str(setosa, 1)

versicolor <- load_as(fs::path(temp_dir, "versicolor.RData"))
str(versicolor, 1)

virginica <- load_as(fs::path(temp_dir, "virginica.RData"))
str(virginica, 1)

# load multiple files in a single R object
loaded_data <- load_multiple(
  files = fs::path(
  temp_dir, c("setosa.RData", "versicolor.RData", "virginica.RData")),
  verbose = TRUE)
str(loaded_data, 1)

# clean up
fs::dir_delete(temp_dir)
}
\author{
Ahmed El-Gabbas
}
