% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_load_as.R
\name{load_as}
\alias{load_as}
\title{Load objects from \code{RData} / \code{qs2} / \code{rds} / \code{feather} file}
\usage{
load_as(file = NULL, n_threads = 1L, timeout = 300L, ...)
}
\arguments{
\item{file}{Character. the file path or URL of the file to be loaded. If
\code{file} is a URL, the function will download the file from the URL to a
temporary file and load it.}

\item{n_threads}{Number of threads to use when reading \code{qs2} files. See
\link[qs2:qs_read]{qs2::qs_read}.}

\item{timeout}{integer; time in seconds before the download times out.
Default 300 seconds; see \link{download.file}.}

\item{...}{Additional arguments to be passed to the respective load
functions. \link[base:load]{base::load} for \code{RData} files; \link[qs2:qs_read]{qs2::qs_read} for \code{qs2} files;
\link[arrow:read_feather]{arrow::read_feather} for \code{feather} files; and \link[base:readRDS]{base::readRDS} for \code{rds}
files.}
}
\value{
Depending on the contents of the \code{RData} file, this function returns
either a single R object or a named list of R objects. The names of the
list elements (if a list is returned) correspond to the names of the
objects stored within the \code{RData} file.
}
\description{
This function loads an \code{RData} file specified by the \code{file} parameter. If the
\code{RData} file contains a single object, that object is returned directly. If
the file contains multiple objects, they are returned as a list with each
object accessible by its name. This allows for flexible handling of loaded
data without needing to know the names of the objects stored within the RData
file ahead of time. The function also supports loading \code{feather}, \code{qs2} and
\code{rds} files.
}
\examples{

file <- system.file("testdata", "culcita_dat.RData", package = "lme4")

# ---------------------------------------------------------
# loading RData using base::load
# ---------------------------------------------------------
(load(file))

ls()

tibble::tibble(culcita_dat)

# ---------------------------------------------------------
# Loading as custom object name
# ---------------------------------------------------------
NewObj <- load_as(file = file)

ls()

print(tibble::tibble(NewObj))

# ---------------------------------------------------------
# Loading multiple objects stored in single RData file
# ---------------------------------------------------------

# store three objects to single RData file
mtcars2 <- mtcars3 <- mtcars

# save in the order of mtcars2, mtcars3, mtcars
TempFile_1 <- tempfile(pattern = "mtcars_", fileext = ".RData")
save(mtcars2, mtcars3, mtcars, file = TempFile_1)

# save in another order: mtcars, mtcars2, mtcars3
TempFile_2 <- tempfile(pattern = "mtcars_", fileext = ".RData")
save(mtcars, mtcars2, mtcars3, file = TempFile_2)


# loading as a single list  with 3 items, keeping original order
mtcars_all_1 <- load_as(TempFile_1)
str(mtcars_all_1, 1)

mtcars_all_2 <- load_as(TempFile_2)
str(mtcars_all_2, 1)
}
\author{
Ahmed El-Gabbas
}
