% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_load_multiple.R
\name{load_multiple}
\alias{load_multiple}
\title{Load multiple \code{.RData} files together}
\usage{
load_multiple(
  files = NULL,
  verbose = TRUE,
  single_object = TRUE,
  return_names = TRUE
)
}
\arguments{
\item{files}{Character vector. The paths of \code{.RData} files to be loaded.}

\item{verbose}{Logical. Whether to print progress messages. Default: \code{TRUE}.}

\item{single_object}{Logical. Whether to load all objects into a single list
(\code{TRUE}) or directly into the global environment (\code{FALSE}). Defaults to
\code{TRUE}.}

\item{return_names}{Logical. Whether to return the names of the loaded
objects. Only effective when \code{single_object} is \code{FALSE}. Defaults to
\code{TRUE}.}
}
\value{
If \code{single_object} is \code{TRUE}, returns a named list of all objects
loaded from the specified files. If \code{single_object} is \code{FALSE} and
\code{return_names} is \code{TRUE}, returns a character vector of the names of the
objects loaded into the global environment. Otherwise, returns \code{NULL}.
}
\description{
This function loads multiple \code{.RData} files either into a single list object
or directly into the global environment. It provides options for verbosity,
returning object names, and handling of non-existent files.
}
\examples{
(files <- system.file(
   "testdata", c("culcita_dat.RData", "gopherdat2.RData"), package = "lme4"))

ls()

# ---------------------------------------------------
# Load multiple *.RData files to one list object
# `single_object = TRUE`
# ---------------------------------------------------
MultiObj <- load_multiple(files = files, single_object = TRUE)
ls()

str(MultiObj, 1)

# ---------------------------------------------------
# Load multiple *.RData files separately to the current environment
# `single_object = FALSE`
# ---------------------------------------------------
ls()
rm("MultiObj")
ls()

load_multiple(files = files, single_object = FALSE)

str(Gdat, 1)

str(culcita_dat, 1)

# ---------------------------------------------------
# Load multiple *.RData files, one object already exists
# ---------------------------------------------------
ls()
rm("culcita_dat")
ls()

try(load_multiple(files = files, single_object = FALSE))

}
\author{
Ahmed El-Gabbas
}
