% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load or install multiple R packages}
\usage{
load_packages(
  ...,
  package_list = NULL,
  verbose = FALSE,
  install_missing = TRUE
)
}
\arguments{
\item{...}{Character. Names of the packages to be loaded or installed.}

\item{package_list}{Character vector. An alternative or additional way to
specify package names as a vector.}

\item{verbose}{Logical. If \code{TRUE}, prints the names and versions of the
loaded packages. Defaults to \code{FALSE}.}

\item{install_missing}{Logical. If \code{TRUE} (default), missing packages are
automatically installed and then loaded.}
}
\value{
This function is used for its side effects (loading/installing
packages) and does not return any value.
}
\description{
This function attempts to load multiple R packages specified by the user. If
a package is not installed, the function can optionally install it before
loading. It also provides an option to print the names and versions of the
loaded packages.
}
\examples{
# Currently loaded packages
(P1 <- ecokit::loaded_packages())

load_packages(tidyr)
# Loaded packages after implementing the function
(P2 <- ecokit::loaded_packages())

# Which packages were loaded?
setdiff(P2, P1)

load_packages(terra, verbose = TRUE)
}
\author{
Ahmed El-Gabbas
}
