% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_load_tar_file.R
\name{load_tar_file}
\alias{load_tar_file}
\title{Load a File from a Tar Archive}
\usage{
load_tar_file(tar_file, file_to_extract, load_fun = "ecokit::load_as", ...)
}
\arguments{
\item{tar_file}{Character. Path to the tar archive file.}

\item{file_to_extract}{Character. Path of the file to extract from the tar
archive (can include directories within the archive).}

\item{load_fun}{Either a function or a character string naming a function
(possibly with package namespace like "package::function") to load the
extracted file. Defaults to \link{load_as}.}

\item{...}{Additional arguments passed to the loading function.}
}
\value{
The object returned by the loading function applied to the extracted
file.
}
\description{
This function extracts a specified file from a tar archive to a temporary
directory, loads it using a provided loading function, and returns the loaded
object.
}
\examples{
ecokit::load_packages(terra, stringr, fs)

# example tar file containing 3 files

tif_file <- system.file("ex/elev.tif", package = "terra")
rds_file <- fs::file_temp(ext = ".rds")
ecokit::save_as(mtcars, out_path = rds_file)
csv_file <- tempfile(fileext = ".csv")
write.csv(iris, csv_file, row.names = FALSE)
tmp_tar <- fs::file_temp(ext = ".tar")
file_list <- c(tif_file, csv_file, rds_file)
base_names <- basename(file_list)
dirs <- dirname(file_list)
for (i in seq_along(file_list)) {
tar_flag <- ifelse(i == 1, "c", "r")
  tar_args <- str_glue(
    'tar -{tar_flag}f {shQuote(tmp_tar)} -C {shQuote(dirs[i])} \\\\
    {shQuote(base_names[i])}')
  invisible(system(tar_args))
}

# List contents of the tar file
print(system2("tar", c("-tf", tmp_tar), stdout = TRUE))

# example SpatRaster
load_tar_file(
  tar_file = tmp_tar, file_to_extract = "elev.tif",
  load_fun = "terra::rast")

# example CSV file using readr
load_tar_file(
  tar_file = tmp_tar, file_to_extract = basename(csv_file),
  load_fun = "read.csv") \%>\%
  head()

# example rds file
load_tar_file(
  tar_file = tmp_tar, file_to_extract = basename(rds_file),
  load_fun = readRDS) \%>\%
  head()
}
\author{
Ahmed El-Gabbas
}
