% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_loaded_packages.R
\name{loaded_packages}
\alias{loaded_packages}
\title{List of loaded packages}
\usage{
loaded_packages()
}
\value{
A character vector containing the names of all loaded packages.
}
\description{
This function returns a character vector listing all the packages that are
currently loaded in the R session.
}
\examples{
loaded_packages()

require(tidyterra)
require(lubridate)

loaded_packages()

}
