% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_n_decimals.R
\name{n_decimals}
\alias{n_decimals}
\title{Number of decimal places in a numeric value}
\usage{
n_decimals(x = NULL)
}
\arguments{
\item{x}{Numeric (or character) numeric value.}
}
\value{
An integer representing the number of decimal places in the input
value. If the input value does not have any decimal places, the function
returns 0.
}
\description{
This function calculates the number of decimal places in a numeric value. It
is designed to work with numeric inputs that can be coerced to character
format.
}
\examples{
n_decimals(x = "13.45554545")

# -------------------------------------------

n_decimals(x = 15.01500)

n_decimals(x = '15.01500')

# -------------------------------------------

n_decimals(x = 13.45554545)
}
\author{
Ahmed El-Gabbas
}
