% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_n_unique.R
\name{n_unique}
\alias{n_unique}
\title{Number of unique values for all columns of a data frame}
\source{
The source code of the function was copied from this
\href{https://stackoverflow.com/q/22196078}{stackoverflow} question.
}
\usage{
n_unique(data)
}
\arguments{
\item{data}{A data frame for which the number of unique values per column
will be calculated.}
}
\value{
A data frame with two columns: \code{Variable} and \code{n_unique}. The
Variable column lists the names of the original columns, and the \code{n_unique}
column lists the number of unique values in each column. The result is
sorted by \code{n_unique} in ascending order.
}
\description{
This function calculates the number of unique values for each column in a
given data frame and returns a data frame with two columns: \code{Variable} and
\code{n_unique}. The \code{Variable} column lists the names of the original columns,
and the \code{n_unique} column lists the corresponding number of unique values in
each column. The result is sorted by the number of unique values in ascending
order.
}
\examples{
n_unique(mtcars)

n_unique(iris)
}
\author{
Ahmed El-Gabbas
}
