% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_nc_global_attributes.R
\name{nc_global_attributes}
\alias{nc_global_attributes}
\title{Get global attributes for \code{NetCDF} files}
\usage{
nc_global_attributes(nc = NULL)
}
\arguments{
\item{nc}{Character. Path to the \code{NetCDF} file. If
\code{NULL}, the function will stop with an error message.}
}
\value{
A character vector where each element is a global attribute.
}
\description{
This function opens a \code{NetCDF} file, extracts all global attributes, and
returns them as a character vector where each element is an attribute
name-value pair.
}
\examples{
require(ecokit)
ecokit::load_packages(stars, sf, fs)

nc_example_1 <- system.file("nc/sub.nc", package = "stars")
if (fs::file_exists(nc_example_1)) nc_global_attributes(nc = nc_example_1)

nc_example_2 <- system.file("nc/timeseries.nc", package = "stars")
if (fs::file_exists(nc_example_2)) nc_global_attributes(nc = nc_example_2)

nc_example_3 <- system.file("nc/cropped.nc", package = "sf")
if (fs::file_exists(nc_example_3)) nc_global_attributes(nc = nc_example_3)
}
\references{
\href{https://github.com/rspatial/terra/issues/1443}{Click here}
}
