% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_normalize_path.R
\name{normalize_path}
\alias{normalize_path}
\title{Normalise and quote file paths}
\usage{
normalize_path(path, must_work = FALSE)
}
\arguments{
\item{path}{Character vector. file path(s).}

\item{must_work}{Logical; if \code{TRUE}, the function errors for non-existing
paths.}
}
\value{
A character vector of absolute, tidied, and shell-quoted paths.
}
\description{
This function ensures that file paths are expressed in a consistent and
canonical form. It first converts paths to absolute form using
\code{\link[fs:path_math]{fs::path_abs()}}, then tidies them with \code{\link[fs:path_tidy]{fs::path_tidy()}}, and finally quotes
them correctly based on the operating system. By default,
\code{\link[base:normalizePath]{base::normalizePath()}} behaves differently on Windows and Linux when a file
does not exist. On Windows, it tries to construct an absolute path, while on
Linux, it returns the input path as-is (relative). To maintain consistency
across platforms, this function uses \code{\link[fs:path_math]{fs::path_abs()}} instead of
\code{\link[base:normalizePath]{base::normalizePath()}}.
}
\author{
Ahmed El-Gabbas
}
