% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_OS.R
\name{OS}
\alias{OS}
\title{Current operating system}
\usage{
OS()
}
\value{
A character string representing the name of the operating system.
}
\description{
This function returns the name of the current operating system the R session
is running on.
}
\examples{
OS()

}
\author{
Ahmed El-Gabbas
}
