% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_package_functions.R
\name{package_functions}
\alias{package_functions}
\title{List of functions in a package}
\usage{
package_functions(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
A character vector containing the names of all functions in the
specified package.
}
\description{
This function returns a character vector listing all the functions available
in the specified R package. It first checks if the package is installed and
can be loaded; if not, it raises an error.
}
\examples{
str(package_functions(package = "raster"))

str(package_functions(package = "sf"))

package_functions(package = "ecokit")

# Error: package not found
 try(package_functions(package = "non_exist"))
}
\author{
Ahmed El-Gabbas
}
