% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_parent_dir.R
\name{parent_dir}
\alias{parent_dir}
\title{Jump Up Parent Directories}
\usage{
parent_dir(
  path = NULL,
  levels = 1L,
  check_dir = FALSE,
  extract_full = FALSE,
  ...
)
}
\arguments{
\item{path}{Character. The file or directory path to start from. If
\code{check_dir = TRUE} and \code{path} exists as a file or has a file extension, its
parent directory is used as the starting point. If \code{check_dir = FALSE}, a
file path is treated as a directory name.}

\item{levels}{Integer. Number of parent levels to jump up. Must be a single,
non-negative integer.}

\item{check_dir}{Logical. If \code{TRUE}, checks \code{path} is a valid directory and
uses its parent if \code{path} is a file. If \code{FALSE}, does not check existence
and uses the input as-is.}

\item{extract_full}{Logical. If \code{TRUE} and \code{levels} exceeds the number of
path components minus one, returns the root of the path. If \code{FALSE}, an
error is thrown when \code{levels} exceeds the available ancestors.}

\item{...}{Additional arguments passed to \link{stop_ctx} for error
reporting.}
}
\value{
Character. The resulting path after jumping. If \code{extract_full = TRUE}
and \code{levels} exceeds the available ancestors, returns the root of the path;
otherwise, returns the ancestor path or throws an error.
}
\description{
Returns the n\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} ancestor directory of a given path.
}
\examples{
parent_dir("/home/user/projects/data/file.txt", levels = 2)
# this will give an error
try(parent_dir("/home/user", levels = 4, extract_full = TRUE))
}
\author{
Ahmed El-Gabbas
}
