% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_polygon_centroid.R
\name{polygon_centroid}
\alias{polygon_centroid}
\title{Replace the geometry of a polygon with its centroid point}
\usage{
polygon_centroid(sf_object = NULL, rename = FALSE, new_name = NULL)
}
\arguments{
\item{sf_object}{A simple feature (\code{sf}) object; the polygon whose geometry
is to be replaced with its centroid. Cannot be \code{NULL}.}

\item{rename}{Logical. Whether to rename the geometry column of the sf
object. Defaults to \code{FALSE}.}

\item{new_name}{Character. New name for the geometry column. Only valid if
\code{rename = TRUE}.}
}
\value{
The modified sf object with its geometry replaced by the centroid of
the original polygon geometry. If rename is \code{TRUE}, the geometry column
will also be renamed as specified by NewName.
}
\description{
This function replaces the geometry of a simple feature (\code{sf}) polygon object
with the geometry of its centroid point. It can optionally rename the
geometry column of the modified \code{sf} object.
}
\references{
\href{https://github.com/r-spatial/sf/issues/480}{Click here}
}
\author{
Ahmed El-Gabbas
}
