% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_range_to_new_value.R
\name{range_to_new_value}
\alias{range_to_new_value}
\title{Changes values within a specified range, or greater than or less than a
specific value to a new value in a vector, data.frame, or raster}
\usage{
range_to_new_value(
  x = NULL,
  between = NULL,
  greater_than = NULL,
  less_than = NULL,
  new_value = NULL,
  invert = FALSE
)
}
\arguments{
\item{x}{A numeric \code{vector}, \code{data.frame}, \code{RasterLayer}, or \code{SpatRaster}
object whose values are to be modified.}

\item{between}{Numeric. A numeric vector of length 2 specifying the range of
values to be changed or kept. If specified, \code{greater_than} and \code{less_than}
are ignored.}

\item{greater_than, less_than}{Numeric. Threshold larger than or equal to/less
than or equal to which values in \code{x} will be changed to \code{new_value}. Only
applied if \code{between} is not specified.}

\item{new_value}{The new value to assign to the selected elements in \code{x}.}

\item{invert}{Logical. Whether to invert the selection specified by
\code{between}. If \code{TRUE}, values outside the specified range are changed to
\code{new_value}. Default is \code{FALSE}.}
}
\value{
The modified object \code{x} with values changed according to the
specified conditions.
}
\description{
This function modifies values in the input object \code{x} based on the specified
conditions. It can operate on vectors, data.frames, or RasterLayer objects.
The function allows for changing values within a specified range (\code{between}),
greater than or equals to (\code{greater_than}) or  less than or equals to
(\code{less_than}) a specified value to a new value (\code{new_value}). An option to
invert the selection is also available for ranges.
}
\examples{
library(raster)
library(terra)
par(mar = c(0.5, 0.5, 1, 2.5), oma = c(0.5, 0.5, 0.5, 1))

# ---------------------------------------------

# Vector

(VV <- seq_len(10))

range_to_new_value(x = VV, between = c(5, 8), new_value = NA)

range_to_new_value(x = VV, between = c(5, 8), new_value = NA, invert = TRUE)

# greater_than is ignored as between is specified
range_to_new_value(
   x = VV, between = c(5, 8), new_value = NA, greater_than = 4)

range_to_new_value(x = VV, new_value = NA, greater_than = 4)

range_to_new_value(x = VV, new_value = NA, less_than = 4)

# ---------------------------------------------

# tibble

iris2 <- iris \%>\%
  tibble::as_tibble() \%>\%
  dplyr::slice_head(n = 50) \%>\%
  dplyr::select(-Sepal.Length, -Petal.Length, -Petal.Width) \%>\%
  dplyr::arrange(-Sepal.Width)

iris2 \%>\%
 dplyr::mutate(
   Sepal.Width.New = range_to_new_value(
      x = Sepal.Width, between = c(3, 3.5),
      new_value = NA, invert = FALSE),
   Sepal.Width.Rev = range_to_new_value(
      x = Sepal.Width, between = c(3, 3.5),
      new_value = NA, invert = TRUE)) \%>\%
 print(n = 50)

# ---------------------------------------------

# RasterLayer / SpatRaster

grd_file <- system.file("external/test.grd", package = "raster")
R_raster <- raster::raster(grd_file)
R_terra <- terra::rast(grd_file)

# Convert values less than 500 to NA
R_raster2 <- range_to_new_value(
x = R_raster, less_than = 500, new_value = NA)
plot(
   raster::stack(R_raster, R_raster2), nr = 1,
   main = c("\nOriginal", "\n<500 to NA"),
   box = FALSE, axes = FALSE, legend.width = 2, colNA = "lightgrey",
   xaxs = "i", yaxs = "i")

R_terra2 <- range_to_new_value(x = R_terra, less_than = 500, new_value = NA)
plot(
   c(R_terra, R_terra2), nr = 1, main = c("\nOriginal", "\n<500 to NA"),
   box = FALSE, axes = FALSE, colNA = "lightgrey", xaxs = "i", yaxs = "i")


# Convert values greater than 700 to NA
R_raster2 <- range_to_new_value(
   x = R_raster, greater_than = 700, new_value = NA)
plot(
   raster::stack(R_raster, R_raster2), nr = 1,
   main = c("\nOriginal", "\n>700 to NA"),
   box = FALSE, axes = FALSE, legend.width = 2, colNA = "lightgrey",
   xaxs = "i", yaxs = "i")

R_terra2 <- range_to_new_value(
    x = R_terra, greater_than = 700, new_value = NA)
plot(
   c(R_terra, R_terra2), nr = 1, main = c("\nOriginal", "\n>700 to NA"),
   box = FALSE, axes = FALSE, colNA = "lightgrey", xaxs = "i", yaxs = "i")
}
\author{
Ahmed El-Gabbas
}
