% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_reload_package.R
\name{reload_package}
\alias{reload_package}
\title{Reload an R package}
\usage{
reload_package(package)
}
\arguments{
\item{package}{Character. Name of the package to reload.}
}
\value{
The function is used for its side effect of reloading a package
rather than for its return value.
}
\description{
This function reloads a specified R package. If the package is not already
loaded, it attempts to load it. If the package is already loaded, it reloads
the package from its library location. This can be useful during package
development when changes are made to the package code. This function depends
on the \link[devtools:reload]{devtools::reload} function.
}
\examples{
load_packages(sf)

reload_package("sf")

}
\author{
Ahmed El-Gabbas
}
