% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_reload_package.R
\name{reload_package}
\alias{reload_package}
\title{Reload an R package}
\usage{
reload_package(...)
}
\arguments{
\item{...}{Unquoted package names (e.g., \code{sf}, \code{ncdf4}). Must be installed
packages. Multiple packages can be specified.}
}
\value{
Returns \code{invisible(NULL)}. The function is used for its side effect
of reloading a package rather than for its return value.
}
\description{
Reloads one or more specified R packages. If a package is not loaded, it is
loaded; if already loaded, it is detached and reloaded from its library
location.
}
\examples{
library(sf)

# Reloads sf and ncdf4. terra0 does not exist
reload_package(sf, ncdf4, terra0)
}
\author{
Ahmed El-Gabbas
}
