% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_options.R
\name{remove_options}
\alias{remove_options}
\title{Remove Options by Name or Pattern}
\usage{
remove_options(pattern = NULL, regex = FALSE, case_sensitive = FALSE)
}
\arguments{
\item{pattern}{Character vector of option names to remove, or a pattern
string if \code{regex = TRUE.}}

\item{regex}{Logical. If \code{TRUE}, treat \code{pattern} as a regular expression to
match option names. Default is \code{FALSE}.}

\item{case_sensitive}{Logical. If \code{TRUE}, matching is case-sensitive. Default
is \code{FALSE}.}
}
\value{
Invisibly returns \code{NULL}. Removes specified options if found.
}
\description{
Removes options from the current R session by name or regular expression
pattern.
}
\examples{
 # removes option named "my_option"
 options(my_option = 42)
 ecokit::extract_options("my_option")
 ecokit::remove_options("my_option")
 ecokit::extract_options("my_option")

 options(plot1 = 42, plot2 = "yes", plot_extra = TRUE)
 ecokit::remove_options("plot_", regex = FALSE)
 ecokit::extract_options("plot")
 ecokit::remove_options("plot_", regex = TRUE)
 ecokit::extract_options("plot")
}
\author{
Ahmed El-Gabbas
}
