% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_replace_space.R
\name{replace_space}
\alias{replace_space}
\title{Replace space with underscore in a string}
\usage{
replace_space(x)
}
\arguments{
\item{x}{Character. The string in which spaces will be replaced with
underscores.}
}
\value{
A character string with all spaces replaced by underscores.
}
\description{
A simple wrapper function for \code{stringr::str_replace_all} that replaces all
spaces with underscores. It is useful for formatting strings to be used in
contexts where spaces are not allowed or desired.
}
\examples{
replace_space("Genus species")

replace_space("Genus species subspecies")
}
