% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_save_multiple.R
\name{save_multiple}
\alias{save_multiple}
\title{Save multiple objects to their respective \code{.RData} files}
\usage{
save_multiple(
  variables = NULL,
  out_directory = getwd(),
  overwrite = FALSE,
  prefix = "",
  verbose = FALSE
)
}
\arguments{
\item{variables}{Character vector. Names of the variables to be saved. If
\code{NULL} or any specified variable does not exist in the global environment,
the function will stop with an error.}

\item{out_directory}{Character. Path to the output folder where the \code{.RData}
files will be saved. Defaults to the current working directory.}

\item{overwrite}{Logical. Whether existing \code{.RData} files should be
overwritten. If \code{FALSE} (Default) and files exist, the function will stop
with an error message.}

\item{prefix}{Character. Prefix of each output file name. Useful for
organizing saved files or avoiding name conflicts. Defaults to an empty
string.}

\item{verbose}{Logical. Whether to print a message upon successful saving of
files. Defaults to \code{FALSE}.}
}
\value{
The function is used for its side effect of saving files and does not
return a value.
}
\description{
This function saves specified variables from the global environment to
separate \code{.RData} files. It allows for optional file prefixing and
overwriting of existing files.
}
\examples{
load_packages(fs, purrr)

temp_dir <- fs::path_temp("save_multiple")
fs::dir_create(temp_dir)

# ----------------------------------------------
# Save x1 and x2 to disk
# ----------------------------------------------
x1 = 10
x2 = 20

save_multiple(
  variables = c("x1", "x2"), out_directory = temp_dir, verbose = TRUE)

list.files(path = temp_dir, pattern = "^.+.RData")

(x1Contents <- ecokit::load_as(fs::path(temp_dir, "x1.RData")))
(x2Contents <- ecokit::load_as(fs::path(temp_dir, "x2.RData")))

# ----------------------------------------------
# Use prefix
# ----------------------------------------------
save_multiple(
  variables = c("x1", "x2"), out_directory = temp_dir, prefix = "A_")

list.files(path = temp_dir, pattern = "^.+.RData")

# ----------------------------------------------
# File exists, no save
# ----------------------------------------------
try(save_multiple(variables = c("x1", "x2"), out_directory = temp_dir))

# ----------------------------------------------
# overwrite existing file
# ----------------------------------------------
x1 = 100; x2 = 200; x3 = 300

save_multiple(
  variables = c("x1", "x2", "x3"),
  out_directory = temp_dir, overwrite = TRUE)

(x1Contents <- ecokit::load_as(fs::path(temp_dir, "x1.RData")))
(x2Contents <- ecokit::load_as(fs::path(temp_dir, "x2.RData")))
(x3Contents <- ecokit::load_as(fs::path(temp_dir, "x3.RData")))

# clean up
fs::dir_delete(temp_dir)
}
\author{
Ahmed El-Gabbas
}
