% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_save_session.R
\name{save_session}
\alias{save_session}
\title{Saves all non-function objects from the global environment to an RData file}
\usage{
save_session(out_directory = getwd(), exclude_objects = NULL, prefix = "S")
}
\arguments{
\item{out_directory}{Character. Directory path where the output \code{RData} file
should be saved. Defaults to the current working directory \code{\link[base:getwd]{base::getwd()}}.}

\item{exclude_objects}{Character vector. Object names (as strings) to exclude
from saving.}

\item{prefix}{Character. prefix the saved file name with. Defaults to \code{S}.}
}
\value{
A tibble containing the names and sizes (in MB, rounded to 1 decimal
place) of the saved objects.
}
\description{
This function saves all objects (except functions and specified exclusions)
from the global environment as list items in an \code{RData} file. It also creates
a summary of these objects' sizes in memory.
}
\author{
Ahmed El-Gabbas
}
