% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_save_session_info.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session information to a file}
\usage{
save_session_info(out_directory = getwd(), session_info = NULL, prefix = "S")
}
\arguments{
\item{out_directory}{Character. Directory path where the output file should
be saved. The default is the current working directory (\link[base:getwd]{base::getwd}).}

\item{session_info}{An optional list of objects to include in the session
information output. This is typically the result of a session management
function like \link{save_session}. If provided, details of these
objects (excluding functions and pre-selected objects, with sizes in
megabytes) are appended to the session information file.}

\item{prefix}{Character. Prefix for the output file name. Defaults to \code{S}.}
}
\value{
The primary effect of this function is the side effect of writing
session information to a file.
}
\description{
This function saves the current R session information, including installed
packages, session details, and optionally, information about specific objects
in the session, to a text file.
}
\author{
Ahmed El-Gabbas
}
