% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_set_geometry.R
\name{set_geometry}
\alias{set_geometry}
\title{Set the geometry column of a simple feature (sf) data frame in the pipe
pipeline.}
\usage{
set_geometry(sf_object, geometry_column)
}
\arguments{
\item{sf_object}{simple feature (sf) data frame. This is the data frame whose
geometry column will be set or changed.}

\item{geometry_column}{Character. Name of the geometry column to be used or
set in the \code{sf_object} data frame.}
}
\value{
The modified simple feature (sf) data frame with the updated geometry
column. The function returns the original data frame \code{sf_object} with its
geometry column set to \code{Name}.
}
\description{
Set the geometry column of a simple feature (sf) data frame, making it
particularly useful in data processing pipelines. By specifying the name of
the geometry column, users can ensure that spatial operations utilise the
correct data.
}
\author{
Ahmed El-Gabbas
}
