% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_set_parallel.R
\name{set_parallel}
\alias{set_parallel}
\title{Set up or stop parallel processing plan}
\usage{
set_parallel(
  n_cores = 1L,
  strategy = "multisession",
  stop_cluster = FALSE,
  show_log = TRUE,
  future_max_size = 500L,
  cat_timestamp = FALSE,
  ...
)
}
\arguments{
\item{n_cores}{Integer. Number of cores to use. If \code{NULL}, defaults to
sequential mode. Default is \code{1}.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[=set_parallel]{set_parallel()}} for details.}

\item{stop_cluster}{Logical. If \code{TRUE}, stops any parallel cluster and
resets to sequential mode. If \code{FALSE} (default), sets up a new plan.}

\item{show_log}{Logical. If \code{TRUE} (default), logs messages via
\code{\link[=cat_time]{cat_time()}}.}

\item{future_max_size}{Numeric. Maximum allowed total size (in megabytes) of
global variables identified. See \code{future.globals.maxSize} argument of
\link[future:zzz-future.options]{future::future.options} for more details. Default is \code{500L} for 500 MB.}

\item{cat_timestamp}{logical; whether to include the time in the timestamp.
Default is TRUE. If FALSE, only the text is printed. See
\code{\link[=cat_time]{cat_time()}}.}

\item{...}{Additional arguments to pass to \link{cat_time}.}
}
\description{
Configures parallel processing with \code{\link[future:plan]{future::plan()}} or stops an existing
plan. When stopping, it resets to sequential mode.
}
\examples{
load_packages(future)

# number of workers available
future::nbrOfWorkers()

# ---------------------------------------------
# `multisession`
# ---------------------------------------------

# Prepare working in parallel
set_parallel(n_cores = 2)
future::plan("list")
future::nbrOfWorkers()

# Stopping parallel processing
set_parallel(stop_cluster = TRUE)
future::plan("list")
future::nbrOfWorkers()

# ---------------------------------------------
# `cluster`
# ---------------------------------------------

# Prepare working in parallel
set_parallel(n_cores = 2, strategy = "cluster")
future::plan("list")
future::nbrOfWorkers()

# Stopping parallel processing
set_parallel(stop_cluster = TRUE)
future::plan("list")
future::nbrOfWorkers()

# ---------------------------------------------
# `multicore`
# ---------------------------------------------

# Prepare working in parallel
set_parallel(n_cores = 2, strategy = "multicore")
future::plan("list")
future::nbrOfWorkers()

# Stopping parallel processing
set_parallel(stop_cluster = TRUE)
future::plan("list")
future::nbrOfWorkers()

# ---------------------------------------------
# `sequential`
# ---------------------------------------------

set_parallel(n_cores = 1, strategy = "sequential")
future::nbrOfWorkers()
}
\author{
Ahmed El-Gabbas
}
