% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_set_raster_crs.R
\name{set_raster_crs}
\alias{set_raster_crs}
\title{Set CRS for a SpatRaster in a Pipeline}
\usage{
set_raster_crs(raster = NULL, crs = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object whose CRS is to be set. Cannot be \code{NULL}
or non-\code{SpatRaster}.}

\item{crs}{Character. A valid CRS string (e.g., EPSG code, WKT, or PROJ4) to
set for the \code{raster}.}
}
\value{
The modified \code{SpatRaster} object with the updated CRS.
}
\description{
Sets the coordinate reference system (CRS) of a \code{SpatRaster} object using a
specified CRS string, designed for use in data processing pipelines (e.g.,
with \verb{\%>\%}). Wraps \verb{terra::crs<-} to assign a valid CRS, such as an EPSG
code.
}
\examples{
load_packages(terra)

# Create a sample SpatRaster, with missing CRS
r <- terra::rast(nrows = 10, ncols = 10, vals = 1:100)
terra::crs(r) <- NULL
print(r)

terra::crs(r, describe = TRUE)$code


# Set CRS to EPSG:4326
(r_modified <- set_raster_crs(r, "epsg:4326"))
terra::crs(r_modified, describe = TRUE)$code
}
\author{
Ahmed El-Gabbas
}
