% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_sf_add_coords.R
\name{sf_add_coords}
\alias{sf_add_coords}
\title{Add longitude and latitude coordinates to an sf object}
\usage{
sf_add_coords(sf_object, name_x = "Long", name_y = "Lat", overwrite = FALSE)
}
\arguments{
\item{sf_object}{An \code{sf} object to which longitude and latitude columns will
be added.}

\item{name_x, name_y}{Character. Name of the longitude column to be added.
Defaults to \code{Long} and \code{Lat}.}

\item{overwrite}{Logical. Whether to overwrite existing columns with names
specified by \code{name_x} and \code{name_y}. If \code{FALSE} and columns with these names
exist, new columns are appended with "_NEW" suffix. Defaults to \code{FALSE}.}
}
\value{
An sf object with added longitude and latitude columns.
}
\description{
Add longitude and latitude coordinates as new columns to an sf object
(\code{sf_object}). It extracts the coordinates from the sf object, converts them
into a tibble, and appends them to the original sf object as new columns. If
\code{name_x} or \code{name_y}, provided as arguments respectively, already exist in
the sf object, the function either  1) overwrites these columns if
\code{overwrite} is set to \code{TRUE} or 2) appends "_NEW" to the new column names to
avoid overwrite if \code{overwrite} is set to \code{FALSE}.
}
\note{
If the overwrite parameter is \code{FALSE} (default) and columns with the
specified names already exist, the function will issue a warning and append
"_NEW" to the names of the new columns to avoid overwriting.
}
\examples{
pt1 = sf::st_point(c(0,1))
pt2 = sf::st_point(c(1,1))
d = data.frame(a = c(1, 2))
d$geom = sf::st_sfc(pt1, pt2)
df = sf::st_as_sf(d)
df
(df <- sf_add_coords(df))

(sf_add_coords(df))

(sf_add_coords(df, overwrite = TRUE))
}
\author{
Ahmed El-Gabbas
}
