% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_source_silent.R
\name{source_silent}
\alias{source_silent}
\title{Silently source R script with optional message and warning suppression}
\usage{
source_silent(file, messages = TRUE, warnings = TRUE, ...)
}
\arguments{
\item{file}{Character. Path of the file to be sourced.}

\item{messages}{Logical. If \code{TRUE} (default), messages generated by the
script are shown. If \code{FALSE}, messages are suppressed.}

\item{warnings}{Logical. If \code{TRUE} (default), warnings generated by the
script are shown. If \code{FALSE}, warnings are suppressed.}

\item{...}{Additional arguments passed to the \link{source} function.}
}
\value{
It is used for its side effect of sourcing a file with optional
suppression of messages and warnings.
}
\description{
This function sources an R script file with options to suppress messages
and/or warnings. It is useful for running scripts that generate a lot of
console output that is not needed.
}
\author{
Ahmed El-Gabbas
}
