% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_split_raster.R
\name{split_raster}
\alias{split_raster}
\title{Split a raster object into a list of smaller rasters}
\usage{
split_raster(
  raster = NULL,
  n_columns = 4L,
  n_rows = 4L,
  save = FALSE,
  save_path = "",
  plot = FALSE,
  extent_only = FALSE
)
}
\arguments{
\item{raster}{A raster object to be split. If NULL (the default), the
function will not execute.}

\item{n_columns, n_rows}{Integer. The desired number of columns and rows to
split the raster into. Default is 4 columns and 4 rows.}

\item{save}{Logical. Whether to save the split rasters to disk. Default is
\code{FALSE}.}

\item{save_path}{Character. Directory path where the split rasters should be
saved if \code{save} is \code{TRUE}. If the directory does not exist, it will be
created.}

\item{plot}{Logical. Whether to plot the split rasters. Default is \code{FALSE}.}

\item{extent_only}{Logical. If \code{TRUE}, the function returns only the extents
of the split rasters instead of the raster data. Default is \code{FALSE}.}
}
\value{
A list of raster objects or extents of the split rasters, depending
on the value of the \code{extent_only} parameter.
}
\description{
Split a raster object into a list of smaller rasters based on specified
numbers of rows and columns. It can optionally save the resulting rasters to
disk, plot them, or return just their extents.
}
\examples{
library(raster)
logo <- raster(system.file("external/rlogo.grd", package = "raster"))
plot(logo, axes = FALSE, legend = FALSE, bty = "n",
     box = FALSE, main = "Original raster layer")
# --------------------------------------------------

# Split into 3 rows and 3 columns
logoSplit <- split_raster(
    raster = logo, n_columns = 3, n_rows = 3, plot = TRUE)

print(logoSplit) # a list object of 9 items

# --------------------------------------------------

# Merging split maps again
logoSplit$fun <- mean
logoSplit$na.rm <- TRUE
logoSplit2 <- do.call(mosaic, logoSplit)
par(mfrow = c(1, 1))
plot(logoSplit2, axes = FALSE, legend = FALSE, bty = "n",
     box = FALSE, main = "Merged raster layers")

print({logoSplit2 - logo}) # No value difference!

# --------------------------------------------------

logoSplit <- split_raster(
   logo, n_columns = 3, n_rows = 3, extent_only = TRUE)
print(logoSplit)

}
\references{
Click \href{https://stackoverflow.com/questions/29784829/}{here} and
\href{https://stackoverflow.com/q/22109774}{here}
}
\author{
Ahmed El-Gabbas
}
