% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_text_to_coordinates.R
\name{text_to_coordinates}
\alias{text_to_coordinates}
\title{Extract longitude and latitude from string}
\usage{
text_to_coordinates(text = NULL, name_x = "Longitude", name_y = "Latitude")
}
\arguments{
\item{text}{Character. Coordinates in the format \code{"POINT (longitude latitude)"}. This parameter is required and cannot be \code{NULL}.}

\item{name_x, name_y}{Character. Name to be used for the longitude and
Longitude columns in the output tibble. Defaults to "Longitude" and
"Latitude".}
}
\value{
A tibble with two columns containing the longitude and latitude
values extracted from the input string. The names of these columns are
determined by the \code{name_x} and \code{name_y} parameters. If no names are
provided, the default names ("Longitude" and "Latitude") are used.

two column tibble for Longitude & Latitude
}
\description{
Extract longitude and latitude from string representing a geographical point
in the format \code{"POINT (longitude latitude)"} and converts it into a
two-column tibble containing the longitude and latitude as numeric values.
The names of the columns in the resulting tibble can be customized. The
default names for the longitude and latitude columns are "Longitude" and
"Latitude", respectively.
}
\examples{
c("POINT (11.761 46.286)", "POINT (14.8336 42.0422)",
  "POINT (16.179999 38.427214)") \%>\%
 lapply(text_to_coordinates)

c("POINT (11.761 46.286)", "POINT (14.8336 42.0422)",
  "POINT (16.179999 38.427214)") \%>\%
 lapply(text_to_coordinates, name_x = "Long", name_y = "Lat")
}
\author{
Ahmed El-Gabbas
}
