% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_zenodo_tools.R
\name{zenodo_tools}
\alias{zenodo_tools}
\alias{zenodo_file_list}
\alias{download_zenodo_file}
\title{List and Download Files from a Zenodo Record}
\usage{
zenodo_file_list(record_id)

download_zenodo_file(
  record_id = NULL,
  zenodo_file = NULL,
  dest_file = NULL,
  read_func = NULL,
  ...
)
}
\arguments{
\item{record_id}{Character or numeric. The Zenodo record ID (e.g.,
\code{"1234567"}).}

\item{zenodo_file}{Character. The file key (or filename) as listed in the
Zenodo record.}

\item{dest_file}{Character (optional). Local destination path to save the
downloaded file. If \code{NULL} (default), a temporary file is created and
deleted if \code{read_func} is used.}

\item{read_func}{Function (optional). A function to read or process the
downloaded file (e.g., \code{\link[=read.csv]{read.csv()}}, \code{\link[readr:read_delim]{readr::read_csv()}}, \code{\link[=readLines]{readLines()}},
\code{\link[readxl:read_excel]{readxl::read_excel()}}, etc.). If \code{NULL}, only the file path is returned.}

\item{...}{Additional arguments passed to \code{read_func}.}
}
\value{
\itemize{
\item \code{zenodo_file_list} returns a \code{tibble} with metadata for each file in the
Zenodo record, including file key, filename, download link, and record-level
metadata such as title, DOI, and creation dates.
\item \code{download_zenodo_file}: If \code{read_func} is \code{NULL}, returns the path to the
downloaded file. If \code{read_func} is provided, returns the output of that
function.
}
}
\description{
These functions provide an R interface to list all files in a Zenodo record
and to download one of those files (optionally reading it into R).
}
\details{
\itemize{
\item \code{zenodo_file_list()} lists all files and their metadata for a given
Zenodo record.
\item \code{download_zenodo_file()} downloads a specified file from a
Zenodo record and can optionally read its contents into R using a
user-supplied function.
}
}
\examples{
require(ecokit)
require(terra)

files <- zenodo_file_list("6620748")
dplyr::glimpse(files)

# --------------------------------------------

# Download as file only
pdf_file <- download_zenodo_file(
   record_id = "1234567", zenodo_file = "article.pdf")
print(pdf_file)

ecokit::file_type(pdf_file)
fs::file_delete(pdf_file)

# Download and read as data frame
nc_file <- download_zenodo_file(
  record_id = "6620748",
  zenodo_file = "100Ma_Pohletal2022_DIB_PhaneroContinentalClimate.nc",
  read_func = terra::rast)

print(class(nc_file))

print(nc_file)
try(fs::file_delete(nc_file), silent = TRUE)
}
\author{
Ahmed El-Gabbas
}
